+++ config/connection.js
@@ -3,17 +3,17 @@ require('dotenv').config();
 
 let sequelize;
 
-console.log(process.env.DB_NAME)
-console.log(process.env.DB_USER)
-console.log(process.env.DB_PASSWORD)
+// console.log(process.env.DB_NAME)
+// console.log(process.env.DB_USER)
+// console.log(process.env.DB_PASSWORD)
 
 if (process.env.JAWSDB_URL) {
   sequelize = new Sequelize(process.env.JAWSDB_URL);
 } else {
   sequelize = new Sequelize(
-    "game_db",
-    "root",
-    "root",
+    process.env.DB_NAME,
+    process.env.DB_USER,
+    process.env.DB_PASSWORD,
     {
       host: 'localhost',
       dialect: 'mysql',
