+++ controllers/homeRoutes.js
@@ -1,6 +1,6 @@
 //These are all the view routes for your application
 const router = require('express').Router();
-const { genMap2, genMapv2 } = require("../models/maps");
+const { genMap2, genMapv2, getRandomMap } = require("../utils/maps");
 //when a GET request is received on the root(/) route,
 //render the home.handlebars view
 router.get('/', (req, res) => {
@@ -12,10 +12,11 @@ router.get('/otherpage', (req, res) => {
   res.render('otherpage');
 });
 
-router.get("/map", (req, res) => {
-  const mapData = genMapv2().map(row => ({ cols: row }));
-  console.log(mapData)
-  res.render("map", { row: mapData })
+router.get("/game", async (req, res) => {
+  // const mapData = genMapv2().map(row => ({ cols: row }));
+  // console.log(mapData)
+  const map = await getRandomMap();
+  res.render("game", map)
 })
 
 module.exports = router;
