+++ package.json
@@ -28,7 +28,8 @@
     "express-handlebars": "^6.0.6",
     "express-session": "^1.17.3",
     "mysql2": "^2.3.3",
-    "sequelize": "^5"
+    "namor": "^2.0.3",
+    "sequelize": "^6.19.0"
   },
   "devDependencies": {
     "dungeon-generator": "^0.0.1",
