+++ views/game.handlebars
@@ -12,7 +12,8 @@
         background: red;
     }
 </style>
-{{#each row}}
+<h1>NAME OF YOUR MAP --- {{name}}</h1>
+{{#each map}}
 <div class="row">
     {{#each cols as |col|}}
         {{#if col}}
